# openapi.committed.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Trading journal REST API",
    "title": "Zorivest API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/analytics/ai-review": {
      "post": {
        "description": "Multi-persona AI trade review. Opt-in with budget cap.",
        "operationId": "ai_review_trade_api_v1_analytics_ai_review_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Ai Review Trade Api V1 Analytics Ai Review Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ai Review Trade",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/v1/analytics/cost-of-free": {
      "get": {
        "description": "'Cost of Free' report \u2014 hidden costs estimate.",
        "operationId": "get_cost_of_free_api_v1_analytics_cost_of_free_get",
        "parameters": [
          {
            "in": "query",
            "name": "account_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Id"
            }
          },
          {
            "in": "query",
            "name": "period",
            "required": false,
            "schema": {
              "default": "ytd",
              "title": "Period",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Cost Of Free Api V1 Analytics Cost Of Free Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Cost Of Free",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/v1/analytics/drawdown": {
      "get": {
        "description": "Monte Carlo drawdown probability table.",
        "operationId": "get_drawdown_table_api_v1_analytics_drawdown_get",
        "parameters": [
          {
            "in": "query",
            "name": "account_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Id"
            }
          },
          {
            "in": "query",
            "name": "simulations",
            "required": false,
            "schema": {
              "default": 10000,
              "title": "Simulations",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Drawdown Table Api V1 Analytics Drawdown Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Drawdown Table",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/v1/analytics/excursion/{trade_exec_id}": {
      "post": {
        "description": "Compute MFE/MAE/BSO metrics for a trade.",
        "operationId": "enrich_trade_excursion_api_v1_analytics_excursion__trade_exec_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "trade_exec_id",
            "required": true,
            "schema": {
              "title": "Trade Exec Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Enrich Trade Excursion Api V1 Analytics Excursion  Trade Exec Id  Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enrich Trade Excursion",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/v1/analytics/execution-quality": {
      "get": {
        "description": "Execution quality scores.",
        "operationId": "get_execution_quality_api_v1_analytics_execution_quality_get",
        "parameters": [
          {
            "in": "query",
            "name": "trade_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Trade Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Execution Quality Api V1 Analytics Execution Quality Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Execution Quality",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/v1/analytics/expectancy": {
      "get": {
        "description": "Win rate, avg win/loss, expectancy per trade, Kelly %.",
        "operationId": "get_expectancy_api_v1_analytics_expectancy_get",
        "parameters": [
          {
            "in": "query",
            "name": "account_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Id"
            }
          },
          {
            "in": "query",
            "name": "period",
            "required": false,
            "schema": {
              "default": "all",
              "title": "Period",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Expectancy Api V1 Analytics Expectancy Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Expectancy",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/v1/analytics/options-strategy": {
      "post": {
        "description": "Auto-detect multi-leg options strategy type.",
        "operationId": "detect_options_strategy_api_v1_analytics_options_strategy_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Detect Options Strategy Api V1 Analytics Options Strategy Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Detect Options Strategy",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/v1/analytics/pfof-report": {
      "get": {
        "description": "PFOF impact analysis report \u2014 labeled as ESTIMATE.",
        "operationId": "get_pfof_report_api_v1_analytics_pfof_report_get",
        "parameters": [
          {
            "in": "query",
            "name": "account_id",
            "required": true,
            "schema": {
              "title": "Account Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "period",
            "required": false,
            "schema": {
              "default": "ytd",
              "title": "Period",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Pfof Report Api V1 Analytics Pfof Report Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Pfof Report",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/v1/analytics/sqn": {
      "get": {
        "description": "System Quality Number (SQN).",
        "operationId": "get_sqn_api_v1_analytics_sqn_get",
        "parameters": [
          {
            "in": "query",
            "name": "account_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Id"
            }
          },
          {
            "in": "query",
            "name": "period",
            "required": false,
            "schema": {
              "default": "all",
              "title": "Period",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Sqn Api V1 Analytics Sqn Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Sqn",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/v1/analytics/strategy-breakdown": {
      "get": {
        "description": "P&L breakdown by strategy name.",
        "operationId": "get_strategy_breakdown_api_v1_analytics_strategy_breakdown_get",
        "parameters": [
          {
            "in": "query",
            "name": "account_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Strategy Breakdown Api V1 Analytics Strategy Breakdown Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Strategy Breakdown",
        "tags": [
          "analytics"
        ]
      }
    },
...
  },
...
}